Adds minimax function for computer AI
The minimax function will allow the computer to play the best move
possible. In order to get this working, a number of functions had to
be tweaked slightly to get the logic working. The isGameOver function
had to be changed a lot so I could check for theoretical winning
positions without changing unwanted values every time it was called.